backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.3.0
httpx==0.26.0
celery==5.3.6
redis==5.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
requests==2.31.0
python-dotenv==1.0.0
itsdangerous==2.1.2
email-validator==2.1.1
